custom: adjust the caret's height and position
The original caret's height is the same as that of the displayed font.
The fonts used in PenCake are all height-enlarged for the natural line spacing.
For this reason, the caret is too long and ugly.
In this patch, we adjusted the caret's height and its position to be natural.
We need to modify two carets, for empty input and text entered input.